Dev/vvmaksim

+++ app/src/main/kotlin/model/graph/classes/DirectedUnweightedGraph.kt
@@ -4,11 +4,11 @@ import model.graph.abstractClasses.AbstractGraph
 import model.graph.dataClasses.UnweightedEdge
 import model.graph.dataClasses.Vertex
 
-class DirectedUnweightedGraph<V : Comparable<V>> : AbstractGraph<Int, V>() {
+class DirectedUnweightedGraph<V : Comparable<V>> : AbstractGraph<Long, V>() {
     fun addEdge(
         from: Vertex<V>,
         to: Vertex<V>,
-    ): Int? = addEdge(UnweightedEdge(nextEdgeId++, from, to))
+    ): Long? = addEdge(UnweightedEdge(nextEdgeId++, from, to))
 
     override fun getNeighbors(vertex: Vertex<V>): List<Vertex<V>> {
         if (vertex !in _vertices.values) return emptyList()
